+++ src/index.ts
@@ -14,17 +14,20 @@ export {
   type ForceSideTypeUpdate,
   type ForceSideTypeInput,
 } from "./lib/forcesides.js";
-export { Federate, Deployment, FederateItem } from "./lib/deployment.js";
+export {
+  Federate,
+  type FederateType,
+  type FederateTypeInput,
+  Deployment,
+} from "./lib/deployment.js";
 export { Holding, type HoldingType } from "./lib/holdings.js";
 export * from "./lib/enums.js";
 export * from "./lib/types.js";
 export * from "./lib/enumValues.js";
 export {
-  EquipmentItemDisposition,
-  UnitDisposition,
-  type DispositionType,
   MsdlCoordinates,
   type MsdlCoordinatesType,
+  MsdlLocation,
 } from "./lib/geo.js";
 export {
   EquipmentModel,
@@ -39,3 +42,8 @@ export {
   type RectangleAreaType,
 } from "./lib/environment.js";
 export { type UnitEquipmentInterface } from "./lib/common.js";
+export {
+  EquipmentItemDisposition,
+  UnitDisposition,
+} from "./lib/disposition.js";
+export type { DispositionType } from "./lib/disposition.js";